Banco de dados

+++ Cafeteria_Carol/Tela_Gerenciar_Produtos.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using System.Globalization;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -14,31 +15,32 @@ namespace Cafeteria_Carol
         
         public partial class Tela_Gerenciar_Produtos : Form
         {
-            private string connectionString = "Data Source=C:\\Users\\gabri\\source\\repos\\Projeto_Cafeteria\\Cafeteria_Carol\\Banco\\bd_cafeteria.db";
+        private string connectionString = ConfiguracaoBanco.CaminhoBanco; 
 
-            public Tela_Gerenciar_Produtos()
+        public Tela_Gerenciar_Produtos()
             {
                 InitializeComponent();
             }
 
         private void bt_AdicionarProduto_Click(object sender, EventArgs e)
+        {
+            string nome = txtNomeProduto.Text;
+            string descricao = txtDescricaoProduto.Text;
+            decimal preco;
+
+            if (decimal.TryParse(txtPrecoProduto.Text, NumberStyles.Currency, CultureInfo.GetCultureInfo("pt-BR"), out preco))
             {
-                string nome = txtNomeProduto.Text;
-                string descricao = txtDescricaoProduto.Text;
-                decimal preco;
-
-                if (decimal.TryParse(txtPrecoProduto.Text, out preco))
-                {
-                    AdicionarProduto(nome, descricao, preco,  null);
-                    MessageBox.Show("Produto adicionado com sucesso!");
-                }
-                else
-                {
-                    MessageBox.Show("Preço inválido. Certifique-se de inserir um valor numérico válido.");
-                }
+                AdicionarProduto(nome, descricao, preco, null);
             }
+            else
+            {
+                MessageBox.Show("Preço inválido. Certifique-se de inserir um valor numérico válido.");
+            }
+        }
 
-            private void bt_ExcluirProduto_Click(object sender, EventArgs e)
+
+
+        private void bt_ExcluirProduto_Click(object sender, EventArgs e)
             {
                 int produtoId;
                 if (int.TryParse(txtIDProduto.Text, out produtoId))
@@ -52,58 +54,90 @@ namespace Cafeteria_Carol
                 }
             }
 
-            private void AdicionarProduto(string nome, string descricao, decimal preco, byte[] imagem)
+        private void AdicionarProduto(string nome, string descricao, decimal preco, byte[] imagem)
+        {
+            using (SQLiteConnection connection = new SQLiteConnection(connectionString))
             {
-                using (SQLiteConnection connection = new SQLiteConnection(connectionString))
+                try
                 {
-                    try
+                    connection.Open();
+
+                    string sql = "INSERT INTO Cardapio (Nome, Descricao, Preco, Imagem) VALUES (@Nome, @Descricao, @Preco, @Imagem)";
+
+                    using (SQLiteCommand command = new SQLiteCommand(sql, connection))
                     {
-                        connection.Open();
+                        command.Parameters.AddWithValue("@Nome", nome);
+                        command.Parameters.AddWithValue("@Descricao", descricao);
+                        command.Parameters.AddWithValue("@Preco", preco);
+                        command.Parameters.AddWithValue("@Imagem", imagem);
 
-                        string sql = "INSERT INTO Cardapio (Nome, Descricao, Preco, Imagem) VALUES (@Nome, @Descricao, @Preco, @Imagem)";
-
-                        using (SQLiteCommand command = new SQLiteCommand(sql, connection))
-                        {
-                            command.Parameters.AddWithValue("@Nome", nome);
-                            command.Parameters.AddWithValue("@Descricao", descricao);
-                            command.Parameters.AddWithValue("@Preco", preco);
-                            command.Parameters.AddWithValue("@Imagem", imagem);
-                          
                         command.ExecuteNonQuery();
-                        }
                     }
-                    catch (Exception ex)
-                    {
-                        MessageBox.Show("Erro ao adicionar o produto: " + ex.Message);
-                    }
+
+                    MessageBox.Show("Produto adicionado com sucesso!");
+
+                    LimparCampos();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Erro ao adicionar o produto: " + ex.Message);
                 }
             }
+        }
 
-            private void ExcluirProduto(int produtoId)
+        private void LimparCampos()
+        {
+            txtNomeProduto.Text = string.Empty;
+            txtDescricaoProduto.Text = string.Empty;
+            txtPrecoProduto.Text = string.Empty;
+            txtIDProduto.Text = string.Empty;
+        }
+
+        private void ExcluirProduto(int produtoId)
+        {
+            using (SQLiteConnection connection = new SQLiteConnection(connectionString))
             {
-                using (SQLiteConnection connection = new SQLiteConnection(connectionString))
+                try
                 {
-                    try
-                    {
-                        connection.Open();
+                    connection.Open();
 
+                    if (ProdutoExiste(produtoId, connection))
+                    {
                         string sql = "DELETE FROM Cardapio WHERE ID = @ID";
 
                         using (SQLiteCommand command = new SQLiteCommand(sql, connection))
                         {
                             command.Parameters.AddWithValue("@ID", produtoId);
-
                             command.ExecuteNonQuery();
+                            MessageBox.Show("Produto excluído com sucesso!");
                         }
                     }
-                    catch (Exception ex)
+                    else
                     {
-                        MessageBox.Show("Erro ao excluir o produto: " + ex.Message);
+                        MessageBox.Show("O produto com ID " + produtoId + " não existe.");
                     }
                 }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Erro ao excluir o produto: " + ex.Message);
+                }
             }
-       
-    private void txtNomeProduto_TextChanged(object sender, EventArgs e)
+        }
+
+        private bool ProdutoExiste(int produtoId, SQLiteConnection connection)
+        {
+            string sql = "SELECT COUNT(*) FROM Cardapio WHERE ID = @ID";
+
+            using (SQLiteCommand command = new SQLiteCommand(sql, connection))
+            {
+                command.Parameters.AddWithValue("@ID", produtoId);
+                int count = Convert.ToInt32(command.ExecuteScalar());
+
+                return count > 0;
+            }
+        }
+
+        private void txtNomeProduto_TextChanged(object sender, EventArgs e)
         {
 
         }
@@ -123,6 +157,11 @@ namespace Cafeteria_Carol
         }
 
         private void Fotos_Enviar_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void Tela_Gerenciar_Produtos_Load(object sender, EventArgs e)
         {
 
         }
